function_modules/sprocket-decode/requirements.txt
redis
ffmpeg-python
paramiko==2.11.0
pycparser==2.21
PyNaCl==1.5.0
pysftp==0.2.9